tt reuni

+++ lib/libft/src/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ibarbouc <ibarbouc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/07 21:17:02 by ibarbouc          #+#    #+#             */
-/*   Updated: 2024/11/28 12:24:32 by ibarbouc         ###   ########.fr       */
+/*   Updated: 2025/07/06 17:53:02 by ibarbouc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
